Avoid reducing player's highest-level when replaying a non-MRU level
If the level was not the most recently played, a save is forced so that
it can be marked as most recently played.  However, the logic to force
the save also forced an update of the highest-level field, even if that
reduced it.

- Rename the parameter to clarify its meaning.
- Add comments explaining the logic to force the save.
- Only update the highest-level field when the update would increase the saved value

+++ common/main/playsave.h
@@ -208,7 +208,7 @@ int read_player_file();
 // set a new highest level for player for this mission
 }
 #endif
-void set_highest_level(int levelnum);
+void set_highest_level(uint8_t levelnum);
 
 // gets the player's highest level from the file for this mission
 int get_highest_level(void);
